+++ R/RECOM_AR.R
@@ -29,14 +29,19 @@ BIN_AR <- function(data, parameter = NULL) {
     )
 
   ## additional measures for sorting the rulebase
-  if(p$sort_measure == "cxs") quality(rule_base) <- cbind(quality(rule_base),
-    cxs = quality(rule_base)$confidence * quality(rule_base)$support)
+  if (p$sort_measure == "cxs")
+    quality(rule_base) <- cbind(quality(rule_base),
+                                cxs = quality(rule_base)$confidence * quality(rule_base)$support)
 
-  if(!p$sort_measure %in% names(quality(rule_base))) {
-      quality(rule_base) <- cbind(
-        quality(rule_base), 
-        setNames(as.data.frame(interestMeasure(rule_base, measure = c(p$sort_measure), transactions = data)), p$sort_measure)
-    )
+  if (!p$sort_measure %in% names(quality(rule_base))) {
+    quality(rule_base) <- cbind(quality(rule_base),
+                                structure(as.data.frame(
+                                  interestMeasure(
+                                    rule_base,
+                                    measure = c(p$sort_measure),
+                                    transactions = data
+                                  )
+                                ), name = p$sort_measure))
   }
 
   ## sort rule_base
@@ -77,7 +82,7 @@ BIN_AR <- function(data, parameter = NULL) {
       ar_duplicates <- duplicated(ar_rhs)
       recom_item <- ar_rhs[!ar_duplicates]
       recom_qual <- ar_qualities[!ar_duplicates]
-      
+
       reclist[[i]] <- if(!is.null(recom_item)) recom_item else integer(0)
       ratings[[i]] <- if(!is.null(recom_qual)) recom_qual else integer(0)
     }